+++ CHANGELOG.md
@@ -30,6 +30,7 @@ after starting a ytdlp download
 - Added `scroll_amount` config option
 - Added `Added()` and `LastModified()` song properties
 - Added `NotExact` and `NotRegex` search options
+- Added `--clean` aragument
 
 ### Changed